moved loadNewsServicesToAnalyze to its own util method
